Add ThreadingMixIn so that the standard 'runserver' command is multithreaded.
Root cause here is an issue with Chrome where it makes multiple
speculative connections to target HTTP servers.  It then chooses one
of these speculative connections to write the request to.  Thus, in a
single-threaded server environment, if the speculative connection
chosen for the "real" request isn't the one that's currently blocked,
then we have to wait for it to timeout, which takes 60s.  This can
make development server sites very slow when using Chrome.  This issue
is particularly visible when using Chrome's Incognito Mode, which
seems to tickle request ordering in exactly the wrong way.